refactor(providers): 파일 크기 가드 준수 분할 — env-validate·openai-compat-mapping 추출
CI Gate 3 (600줄 파일 가드) 초과 해소:
- env.ts 612→535줄: validateConfig 를 env-validate.ts 로 이동 (기존 경로 재노출)
- openai-compat-provider.ts 625→503줄: 메시지/도구 변환 헬퍼
  (toOpenAIMessages/toOpenAITools/needsExplicitPromptCache)를
  openai-compat-mapping.ts 로 이동. 동작 변경 없음
